+++ at_daemon/at_daemon_core/lib/at_daemon_client_core.dart
@@ -1,3 +1,3 @@
 library at_daemon_client_core;
 
-export 'src/models/session.dart';
+export 'src/session/session.dart';
